Moved all helper functions into more descriptive modules.
