+++ fabfile.py
@@ -1,31 +1,37 @@
 import json
-import sqlite3
 import urllib2
 
-from fabric.api import *
+from fabric.api import task, local
+
 
 @task
 def setup():
     install_deps()
     init_db()
 
+
 @task
 def install_deps():
     local('virtualenv . -p python3')
     local('./bin/pip3 install -r requirements.txt')
 
+
 @task
 def init_db():
-    local('bin/python3 -c "from periodo import init_db; init_db()"')
+    local('bin/python3 -c "from periodo.commands import init_db; init_db()"')
+
 
 @task
 def load_data(datafile):
-    local('./bin/python3 -c "from periodo import load_data; load_data(\'{}\')"'.format(datafile))
+    local('./bin/python3 -c "from periodo.commands import load_data; '
+          + 'load_data(\'{}\')"'.format(datafile))
     print('Data loaded from {}.'.format(datafile))
+
 
 @task
 def test():
     local('bin/python3 -m unittest discover')
+
 
 @task
 def get_latest_client():
@@ -40,7 +46,6 @@ def get_latest_client():
     with open('client.zip', 'w') as tmp_zip:
         tmp_zip.write(zip_resp.read())
 
-    local('mkdir -p static/html')
-    local('unzip client.zip -d static/html')
+    local('mkdir -p periodo/static/html')
+    local('unzip client.zip -d periodo/static/html')
     local('rm client.zip')
-
